Work in progress on converting to unit library and using correct values to control Talon on Elevator. Work in progress on AxisConfiguration for elevator.  Still failing unit test on validation, so there's bugs there.

+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Elevator/ElevatorMoveToPosition.java
@@ -2,15 +2,16 @@ package org.usfirst.frc2813.Robot2018.commands.Elevator;
 
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
+import org.usfirst.frc2813.units.values.Length;
 
 /**
  * Move elevator to position. Motor controller does this without
  * further intervention.
  */
 public class ElevatorMoveToPosition extends GearheadsInstantCommand {
-	private final double position;
+	private final int position;
 
-	public ElevatorMoveToPosition(double position) {
+	public ElevatorMoveToPosition(int position) {
 		this.position = position;
 		requires(Robot.elevator);
 	}
